+++ src/handleSnapshotCheck.ts
